pyproject.toml
[project]
name = "rrrymz"
version = "0.1.0"
description = "AI News Curator – multi-agent pipeline for discovering, verifying, and deduplicating global news"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "trafilatura>=2.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "feedparser>=6.0.0",
    "duckduckgo-search>=7.0.0",
]

[project.scripts]
curator-run = "curator.pipeline:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/curator"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
